Add RecordInfo table, remove explicit RecordClass
RecordInfo enforces consistency between records of a given name pointing
to a given ensemble. This allows for the first record to be processed to
define it for all realisations.

Furthermore, this commit removes explicit POSTing of RecordClass as
parameters, as this information is already present when an ensemble is
created.